+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 7.73
 -----
-
+*   Updated
+    *   Dark theme improvements on the podcast page
+        ([#2811](https://github.com/Automattic/pocket-casts-android/pull/2811))
 
 7.72
 -----